SCB-2094 Add server metrics pkg
