+++ datasource/etcd/event/service_event_handler.go
@@ -26,7 +26,7 @@ import (
 	"github.com/apache/servicecomb-service-center/pkg/log"
 	pb "github.com/apache/servicecomb-service-center/pkg/registry"
 	"github.com/apache/servicecomb-service-center/server/core"
-	"github.com/apache/servicecomb-service-center/server/service/metrics"
+	"github.com/apache/servicecomb-service-center/server/metrics"
 	"strings"
 )
